Some cleanup-work and reduced the warning count
Former-commit-id: 64e67c03c2076d457e3b3eb9a27a6fed28414b59 [formerly 630cd0aa40bf14201c731413e4f37cf16f7da011]
Former-commit-id: 2c059b9942ccc9d1c0f4902aff24ab7b6b8937e8

+++ fancy_core/GpuBufferDX12.cpp
@@ -102,7 +102,7 @@ namespace Fancy {
     myStorage = nullptr;
   }
 //---------------------------------------------------------------------------//
-  void* GpuBufferDX12::Lock(GpuResoruceLockOption eLockOption, uint uOffsetElements /* = 0u */, uint uNumElements /* = 0u */) const
+  void* GpuBufferDX12::Lock(GpuResoruceLockOption eLockOption, uint64 uOffsetElements /* = 0u */, uint64 uNumElements /* = 0u */) const
   {
     ASSERT(uOffsetElements + uNumElements <= myProperties.myNumElements);
 
@@ -132,7 +132,7 @@ namespace Fancy {
     return mappedData;
   }
 //---------------------------------------------------------------------------//
-  void GpuBufferDX12::Unlock(uint anOffsetElements /* = 0u */, uint aNumElements /* = 0u */) const
+  void GpuBufferDX12::Unlock(uint64 anOffsetElements /* = 0u */, uint64 aNumElements /* = 0u */) const
   {
     if (anOffsetElements == 0u && aNumElements == 0u)
       aNumElements = myProperties.myNumElements;
