+++ fancy_core/GpuBufferDX12.h
@@ -13,8 +13,8 @@ namespace Fancy {
     ~GpuBufferDX12() override;
         
     void Create(const GpuBufferProperties& clParameters, const void* pInitialData = nullptr) override;
-    void* Lock(GpuResoruceLockOption eLockOption, uint uOffsetElements = 0u, uint uNumElements = 0u) const override;
-    void Unlock(uint anOffsetElements = 0u, uint aNumElements = 0u) const override;
+    void* Lock(GpuResoruceLockOption eLockOption, uint64 uOffsetElements = 0u, uint64 uNumElements = 0u) const override;
+    void Unlock(uint64 anOffsetElements = 0u, uint64 aNumElements = 0u) const override;
 
   protected:
     void Destroy();
